chore(phd2-guider): zero the L5 lint census (20 prod sites) (#948)
* chore(phd2-guider): zero the L5 lint census (20 prod sites)

L5l of the workspace-lints plan: as_conversions,
arithmetic_side_effects and indexing_slicing to zero production
sites in phd2-guider — the first rung with no new #[expect] and no
clippy.toml change.

- Slice patterns fold the existing length checks into the access
  (parse_roi, lock position, camera frame size)
- serde_json index-assign -> Map::insert (roi, temperature params)
- u64->u32 casts -> u32::try_from into the existing error chains
- sample_count retyped u32 -> usize end-to-end (wire shape unchanged)
- Counters saturate (reconnect attempt, RMS window ra_n/dec_n)
- Deadlines made total: settle backstop via saturating_add + pinned
  sleep in select!; stop poll via checked_add with far-future None

Verification: three-lint census reads 0 on --lib --bins; bazel gate
99/99; cargo fmt + clippy -D warnings clean; 373 crate tests pass.


* chore(phd2-guider): name the u32 range in the conversion error messages

Copilot round 1 on #948: the try_from rewrite made "Expected
integer…" cover both a non-integer and an integer above u32::MAX.
Saying "unsigned 32-bit integer" makes the one message truthful for
both failure modes (exposure duration, width, height).


* docs(plans): sync the L5l error-message wording with f51f4f31

Copilot round 2 (suppressed) caught the plan doc still quoting the
pre-reword "Expected integer…" message.


* docs(plans): exact-length slice pattern in the L5l example

Copilot round 3 (suppressed): the illustrative pattern said
[x, y, ..], which also matches longer arrays; the code uses exact
patterns, so the example now does too.


---------

+++ services/phd2-guider/src/service/api.rs
@@ -81,7 +81,7 @@ struct SettledResponse {
     rms_ra_px: Option<f64>,
     rms_dec_px: Option<f64>,
     total_rms_px: Option<f64>,
-    sample_count: u32,
+    sample_count: usize,
 }
 
 impl SettledResponse {
@@ -105,7 +105,7 @@ struct StatsResponse {
     total_rms_px: Option<f64>,
     snr: Option<f64>,
     star_mass: Option<f64>,
-    sample_count: u32,
+    sample_count: usize,
 }
 
 #[derive(Debug, Serialize)]
